2.0
Last Checkpoint

+++ Bees/src/main/java/com/bees/bees/Word.java
@@ -10,7 +10,7 @@ public class Word {
     public String letters;
     public int point;
     public boolean isPangram;
-    public int hashCode;
+    public long hashCode;
 
     public boolean checkWord() {
         if(!isTurkish())
